+++ src/backend/joanie/tests/core/test_api_contract_definitions.py
@@ -4,6 +4,8 @@ Test suite for Contract Definition API.
 from http import HTTPStatus
 from io import BytesIO
 
+from django.contrib.sites.models import Site
+
 from pdfminer.high_level import extract_text as pdf_extract_text
 
 from joanie.core import factories
@@ -134,6 +136,13 @@ class ContractDefinitionApiTest(BaseAPITestCase):
         user = factories.UserFactory(
             email="student_do@example.fr", first_name="John Doe", last_name=""
         )
+        factories.SiteConfigFactory(
+            site=Site.objects.get_current(),
+            terms_and_conditions="""
+            ## Terms and conditions
+            Here are the terms and conditions of the current contract
+            """,
+        )
         factories.UserAddressFactory(owner=user)
         contract_definition = factories.ContractDefinitionFactory()
         token = self.get_user_token(user.username)
@@ -158,6 +167,7 @@ class ContractDefinitionApiTest(BaseAPITestCase):
             "This document certifies that the student wants to enroll to the course",
             document_text,
         )
+        self.assertIn("Terms and conditions", document_text)
         self.assertIn("Student's signature", document_text)
         self.assertIn("[SignatureField#1]", document_text)
         self.assertIn("Representative's signature", document_text)